+++ jirust-cli/src/lib.rs
@@ -8,6 +8,7 @@ use clap::Parser;
 use config::config_file::ConfigFile;
 use executors::config_executor::ConfigExecutor;
 use executors::jira_commands_executors::jira_issue_executor::IssueExecutor;
+use executors::jira_commands_executors::jira_issue_link_executor::LinkIssueExecutor;
 use executors::jira_commands_executors::jira_issue_transition_executor::IssueTransitionExecutor;
 use executors::jira_commands_executors::jira_project_executor::ProjectExecutor;
 use executors::jira_commands_executors::ExecJiraCommand;
@@ -147,6 +148,10 @@ pub async fn process_command(
             let issue_transition_executor =
                 IssueTransitionExecutor::new(cfg_data, args.transition_act, args);
             issue_transition_executor.exec_jira_command().await?
+        }
+        Commands::Link(args) => {
+            let link_issue_executor = LinkIssueExecutor::new(cfg_data, args.link_act, args);
+            link_issue_executor.exec_jira_command().await?
         }
     }
     Ok(())