+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 ## [Unreleased]
+### Added
+- Manage multiple possible resolution (or final) transitions on automatic resolution in Jira version creation
 
 ## [0.10.0] 2024-11-21
 ### Added