+++ CHANGELOG.md
@@ -3,13 +3,17 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
+
 ## [Unreleased]
 ### Fixed
 - Fixed the format used in jira_doc_std_field macro to allow to use it properly
 
 ### Added
 - Manage multiple possible resolution (or final) transitions on automatic resolution in Jira version creation
+- Added a new command to link Jira issues
 
+### Changed
+- Updated dependencies
 
 ## [0.10.0] 2024-11-21
 ### Added